feat: ensure DbDeleteQuestion returns true on success

+++ src/data/usecases/delete-question/delete-question.ts
@@ -1,4 +1,5 @@
 import { DeleteQuestion } from '../../../domain/usecases/delete-question'
+import { InvalidParamError } from '../../../presentation/errors/invalid-param-error'
 import { DeleteQuestionRepository } from '../../protocols/delete-question-repository'
 import { GetQuestionRepository } from '../../protocols/get-question-repository'
 
@@ -7,12 +8,15 @@ export class DbDeleteQuestion implements DeleteQuestion {
   private readonly getQuestionRepository: GetQuestionRepository
 
   constructor (deleteQuestionRepository: DeleteQuestionRepository, getQuestionRepository: GetQuestionRepository) {
-    this.deleteQuestionRepository = deleteQuestionRepository
     this.getQuestionRepository = getQuestionRepository
+    this.deleteQuestionRepository = deleteQuestionRepository
   }
 
   async delete (id: string): Promise<Boolean> {
-    await this.getQuestionRepository.get(id)
+    const questionExist = await this.getQuestionRepository.get(id)
+    if (!questionExist) {
+      throw new InvalidParamError('id')
+    }
     await this.deleteQuestionRepository.delete(id)
     return true
   }
